requirements.txt
flask==2.0.3
werkzeug==2.0.3
requests==2.27.1
python-dotenv==0.20.0
flask-cors==3.0.10
py-trello==0.19.0